// media-scraper-api/src/modules/scraper/scraper.enum.ts
export enum ScraperQueueName {
  SCRAPE_URL = 'scrape-url',
}

export enum ScraperQueue {
  SCRAPER = 'scraper',
}

export enum MediaType {
  IMAGE = 'image',
  VIDEO = 'video',
}

export enum ScrapingStatus {
  PENDING = 'pending',
  PROCESSING = 'processing',
  COMPLETED = 'completed',
  FAILED = 'failed',
}

export enum MediaTypeFilter {
  IMAGE = 'image',
  VIDEO = 'video',
  ALL = 'all',
}